refactor(core): extract shared sqlite driver abstraction
